LayoutTests/platform/wpe/imported/w3c/web-platform-tests/css/css-tables/tentative/table-quirks-expected.txt

Tables in quirks mode proposals
The table cell height box sizing quirk

100 height
100 height
The table cell nowrap minimum width calculation quirk

Chrome Legacy, Edge, Safari fail, only FF gets it correct.Proposal: deprecate the quirk

nowrap nowrap nowrap nowrap
The text decoration doesn’t propagate into tables quirk

decoration
The collapsing table quirk

Chrome Legacy/Edge/Safari ignore the quirk, FF does not. Proposal: depreciate the quirk

The table cell width calculation quirk



PASS decoration does not propagate into table
PASS table 1
PASS table 2
FAIL table 3 assert_equals:
<table>
  <tbody><tr>
    <td nowrap="" style="width:50px;font: 20px/1 Ahem" data-expected-width="580">nowrap nowrap nowrap nowrap</td>
  </tr>
</tbody></table>
width expected 580 but got 295
PASS table 4
PASS table 5
PASS table 6
